requirements.txt
opencv-python
numpy<2.0.0
pillow
PyYAML
setuptools
torch
torchvision
tqdm